==> UEMC - Master in AI and Big Data/3. Programming techniques/Activity 1/Proyecto/1.Actividad1-Limpieza.ipynb <==
import pandas as pd
import matplotlib.pyplot as plt

### 1. Lectura de los datos con python. Ficheros de navegación y de conversión

data_location_navegacion = 'navegacion.csv'
data_location_conversiones = 'conversiones.csv'
df_navegacion = pd.read_csv(data_location_navegacion)
df_conversiones = pd.read_csv(data_location_conversiones)

### 2. Separación de los datos en columnas y concatenarlo a navegacion
Obtener por cada línea de navegación: campaña, adgoup, advertisement y site link que se obtiene de la columna URL.

SE HAN OBTENIDO TODOS LOS CAMPOS DE LA URL EN UN DATAFRAME

def separate_data(df_navegacion):
    url_list_withoutnan = eliminate_nanvalues_from_url(df_navegacion)
    columns_list = ['URL_base&landing', 'glid', 'iduser', 'uuid', 'camp', 'adg', 'device','sl', 'adv', 'rec', 'extra']
    df_url = create_url_dataframe(columns_list)
    df_url_completed = get_url_data(df_url, url_list_withoutnan)
    df_url_completed.to_csv("df_url_completed.csv", index = False)
    df_navegacion_url_separated = pd.concat([df_navegacion, df_url_completed], axis = 1)
    df_navegacion_url_separated.to_csv("df_navegacion_url_separated.csv", index = False)
    
    return df_navegacion_url_separated
    
df_navegacion_url_separated = separate_data(df_navegacion)


#### 2.1. Eliminar valores NAN de la columna de url

def eliminate_nanvalues_from_url(df_navegacion):
    url_list = df_navegacion["url_landing"]
    url_list_withoutnan = [x for x in url_list if x != 'NaN']
    df_navegacion["url_landing"] = url_list_withoutnan
    
    return url_list_withoutnan

#### 2.2. Crear un dataframe para dividir los campos de cada url

def create_url_dataframe(columns_list):
    df_url = pd.DataFrame(columns = columns_list)
    return df_url

#### 2.3. Dividir la url e introducir los valores en cada uno de sus campos en la columna correspondiente

def get_url_data(df_url, url_list_withoutnan):
    i = 0
    for url in url_list_withoutnan:
        if (url != "") and (isinstance(url, str)):
            fila = []
            
            #1. Se parte la url por el elemento '&'
            splitted_url = url.split('&')

            # 2. Se toma el primer y el segundo elemento de splitted_url. 
            #La primera parte contiene url_base y landing. La segunda parte gclid
            splitted_url_landing_glid = splitted_url[0].split('gclid=')

            # 3. Se parte de nuevo el primer elemento de splitted_url_landing_glid. 
            # Se parte en url_landing y en glid
            url_landing = splitted_url_landing_glid[0]
            if(len(splitted_url_landing_glid) >= 2):
                glid = splitted_url_landing_glid[1]

            fila = splitted_url
            fila[0] = glid
            fila.insert(0, url_landing)

            if len(fila) < 11:
                fila.append(" ")

            df_url.loc[i] = fila
        i = i + 1
    return df_url

### 3. Identificar usuarios repetidos

df_navegacion_url_separated = pd.read_csv('df_navegacion_url_separated.csv')
usuarios_unicos = df_navegacion_url_separated['id_user'].unique()
len(usuarios_unicos)

df_navegacion_url_separated[df_navegacion_url_separated.id_user.isnull()]


navegacion_url_separated_duplicated = df_navegacion_url_separated['id_user'].duplicated()
print(navegacion_url_separated_duplicated)

usuarios_repetidos = 0
for element in navegacion_url_separated_duplicated:
    if element == True:
        usuarios_repetidos = usuarios_repetidos + 1

print(usuarios_repetidos)

#### 3.1. Orden de los usuarios por la columna ts para quedarnos con los primeros usuarios no repetidos

df_navegacion_url_separated.sort_values(['ts', 'id_user'], ascending=[True, False])

#### 3.2. Eliminar usuarios repetidos

Como ahora tenemos todos los datos ordenados por ts y por usuario y drop_duplicates se queda con el primer usuario de todos los repetidos, nos aseguramos quedarnos con el registro del usuario que accedió antes.

df_navegacion_urlseparated_noduplicates = df_navegacion_url_separated.drop_duplicates(subset=['id_user'])
df_navegacion_urlseparated_noduplicates.to_csv("df_navegacion_urlseparated_noduplicates.csv", index = False)

### 4. Unir tablas
Para ello hay que buscar id_user en conversiones. Si se encuentra se añade un 1 en una nueva columna llamada conversiones y 0 en caso contrario. Además se toman los valores de id_lead, lead_type y result_list

def añadir_columnas_conversion(df_navegacion_urlseparated_noduplicates, df_conversiones_withdates):
    
    conversion_list = []
    id_lead_list = []
    lead_type_list = []
    result_list = []
    hour_list = []
    day_list = []
    month_list = []
    year_list = []
    
    users_conversiones_list = df_conversiones_withdates['id_user'].tolist()
    print(users_conversiones_list)
    
    # Por cada fila de navegacion
    for row in df_navegacion_without_reps_url_separated.itertuples():
        id_user = row.id_user
        
        str_match = [s for s in users_conversiones_list if s.__contains__(id_user)]
        
        # Si hay match
        if str_match != []:
            print(str_match[0])
            
            # Nos recorremos la tabla de conversiones en busca de la fila que contiene el id_user con el que ha hecho match
            for row_conversiones in df_conversiones_withdates.itertuples():
                print(row_conversiones.id_user)
                if str_match[0] == str(row_conversiones.id_user):
                    element_conversion = 1
                    element_id_lead = row_conversiones.id_lead
                    element_lead_type = row_conversiones.lead_type
                    element_list_append = row_conversiones.result
                    element_list_hour = row_conversiones.hour
                    element_list_day = row_conversiones.day
                    element_list_month = row_conversiones.month
                    element_list_year = row_conversiones.year
                    
                    conversion_list.append(element_conversion)               
                    id_lead_list.append(element_id_lead)
                    lead_type_list.append(element_lead_type)
                    result_list.append(element_list_append)
                    hour_list.append(element_list_hour)
                    day_list.append(element_list_day)
                    month_list.append(element_list_month)
                    year_list.append(element_list_year)
                    
        else:
            element_conversion = 0
            element_id_lead = '-'
            element_lead_type = '-'
            element_list_append = '-'
            element_list_hour = '-'
            element_list_day = '-'
            element_list_month = '-'
            element_list_year = '-'
            
            conversion_list.append(element_conversion)
            id_lead_list.append(element_id_lead)
            lead_type_list.append(element_lead_type)
            result_list.append(element_list_append)
            hour_list.append(element_list_hour)
            day_list.append(element_list_day)
            month_list.append(element_list_month)
            year_list.append(element_list_year)

    
    df_navegacion_without_reps_url_separated['conversiones'] = pd.Series(conversion_list)
    df_navegacion_without_reps_url_separated['id_lead'] = pd.Series(id_lead_list)
    df_navegacion_without_reps_url_separated['lead_type'] = pd.Series(lead_type_list)
    df_navegacion_without_reps_url_separated['result'] = pd.Series(result_list)
    df_navegacion_without_reps_url_separated['hour_conversion'] = pd.Series(hour_list)
    df_navegacion_without_reps_url_separated['day_conversion'] = pd.Series(day_list)
    df_navegacion_without_reps_url_separated['month_conversion'] = pd.Series(month_list)
    df_navegacion_without_reps_url_separated['year_conversion'] = pd.Series(year_list)
    
    
    df_navegacion_without_reps_url_separated.to_csv("df_navegacion_urlseparated_noduplicates_conversiones_concatenated.csv", index = False)

df_navegacion_without_reps_url_separated = pd.read_csv("df_navegacion_urlseparated_noduplicates.csv")  
df_conversiones_withdates = pd.read_csv("df_conversiones_withdates.csv")
añadir_columnas_conversion(df_navegacion_without_reps_url_separated, df_conversiones_withdates)

### 5. Separar la columna ts en las columnas de hora, día, mes y año

#### MAIN FUNCTION

def create_customers_conversiones_withconversiones_withtime(dataframe):
    df_dates = separate_ts(dataframe)
    
    dataframe = dataframe.drop(['ts'], axis = 1)
    df_navegacion_without_reps_url_separated_conversiones_concatenated_withtime = pd.concat([df_dates, dataframe], axis = 1)
    df_navegacion_without_reps_url_separated_conversiones_concatenated_withtime.to_csv('df_navegacion_urlseparated_noduplicates_conversiones_concatenated_withtime.csv', index = False)
    
df_navegacion_without_reps_url_separated_conversiones_concatenated = pd.read_csv("df_navegacion_urlseparated_noduplicates_conversiones_concatenated.csv")  
create_customers_conversiones_withconversiones_withtime(df_navegacion_without_reps_url_separated_conversiones_concatenated)

def separate_ts(dataframe):
    ts = dataframe["ts"]
    # Guardar en un Dataframe la fecha (Día, mes, año)
    columns_list = ["day", "month", "year", "hour"]
    df_dates = pd.DataFrame(columns = columns_list)
    print(df_dates)

    i = 1
    for t in ts:
        df_dates_row = []
        splitted_t = t.split(' ')
        fecha = splitted_t[0]
        hour = splitted_t[1]
        splitted_fecha = fecha.split('-')
        day = splitted_fecha[2]
        month = splitted_fecha[1]
        year = splitted_fecha[0]

        df_dates_row.append(day)
        df_dates_row.append(month)
        df_dates_row.append(year)
        df_dates_row.append(hour)

        print(df_dates_row)

        df_dates.loc[len(df_dates.index)] = df_dates_row

        i = i + 1

    return df_dates

Separación de la fecha en conversiones.csv

# Guardar en un Dataframe la fecha (Día, mes, año)
columns_list = ["day", "month", "year"]
df_dates = pd.DataFrame(columns = columns_list)
dates = df_conversiones["date"]
print(df_dates)

for date in dates:
    splitted_date = date.split('/')
    df_dates.loc[len(df_dates.index)] = splitted_date
    
df_conversiones_withdates = pd.concat([df_dates, df_conversiones], axis = 1)
df_conversiones_withdates.to_csv('df_conversiones_withdates.csv', index = False)
