==> UEMC - Master in AI and Big Data/3. Programming techniques/Activity 1/Proyecto/2.Actividad1-Informe.ipynb <==
import pandas as pd
from datetime import datetime

**Primero de todo hacer una copia de la tabla de la url separada en un CSV para poder trabajar con él**

clean_dataframe = pd.read_csv("df_navegacion_urlseparated_noduplicates_conversiones_concatenated_withtime.csv")

## 1. Informe de ratio de conversión


def get_campañas_data():
    campañas_list = clean_dataframe['camp'].unique()

    columns_list = ["campaña", "numero_conversiones", "numero_clicks", "ratio_conversion"]
    df_campañas_data = pd.DataFrame(columns = columns_list)

    campaña_list = []
    numero_conversiones_list = []
    numero_clicks_list = []
    ratio_conversion_list = []

    for campaña in campañas_list:
        if 'camp' in str(campaña):
            df_campaña = clean_dataframe.loc[clean_dataframe['camp'] == campaña]
            
            campaña_list.append(campaña)

            numero_conversiones = len([df_campaña['conversiones'] == 1])
            numero_conversiones_list.append(numero_conversiones)

            df_campaña = clean_dataframe.loc[clean_dataframe['camp'] == campaña]
            numero_clicks = len(df_campaña)
            numero_clicks_list.append(numero_clicks)

            ratio_conversion = numero_conversiones/numero_clicks * 100
            ratio_conversion_list.append(ratio_conversion)

    df_campañas_data['campaña'] = campaña_list
    df_campañas_data['numero_conversiones'] = numero_conversiones_list
    df_campañas_data['numero_clicks'] = numero_clicks_list
    df_campañas_data['ratio_conversion'] = ratio_conversion_list
    
    df_campañas_data.to_csv('ratio_conversion_campañas_data.csv', index = False)
    
    return df_campañas_data

                    
df_campañas_data = get_campañas_data()
df_campañas_data

def get_informe_ratio_conversion():
    # Cálculo del ratio de conversión
    df_only_conversiones = clean_dataframe.loc[clean_dataframe["conversiones"] == 1]
    numero_clientes_conversion = len(df_only_conversiones)
    numero_clientes_total = len(clean_dataframe)
    ratio_conversion = numero_clientes_conversion/numero_clientes_total
    
    # Guardado de los registros que contienen conversiones = 1 (es decir que han convertido)
    df_only_conversiones.to_csv("df_only_conversiones.csv")
    
    return ratio_conversion
                    
ratio_conversion = get_informe_ratio_conversion()
print("El ratio de conversión es: " + str(ratio_conversion))

## 2. Media de tiempo que tarda el call center en ponerse en contacto con el usuario cuando rellena un formulario

#### 2.1. Recoger todos los usuarios que tienen FORM en lead_type

call_clean_dataframe = clean_dataframe.loc[clean_dataframe['lead_type']==' FORM']
call_clean_dataframe.to_csv('CALL_contact.csv', index = False)
#call_clean_dataframe.drop(call_clean_dataframe.index[13])

def get_media_tiempo(dataframe):
    diferencia_horas_list = []
    for element in dataframe.itertuples():

        # Tiempo de la navegación
        navegacion_day = element.day
        navegacion_hour = element.hour.strip()
        splitted_navegacion_hour = navegacion_hour.split(':')
        navegacion_horas = int(splitted_navegacion_hour[0])
        navegacion_minutos = int(splitted_navegacion_hour[1])

        # Tiempo de la navegación en minutos
        time_navegacion_minutos = navegacion_day*24*60 + navegacion_horas*60 + navegacion_minutos

        # Tiempo de la conversión
        conversion_day = int(element.day_conversion.strip())
        conversion_hour = element.hour_conversion.strip()
        splitted_conversion_hour = conversion_hour.split(':')
        conversion_horas = int(splitted_conversion_hour[0])
        conversion_minutos = int(splitted_conversion_hour[1])

         # Tiempo de la conversión en minutos
        time_conversion_minutos = conversion_day*24*60 + conversion_horas*60 + conversion_minutos
        
        # Cálculo del tiempo que el call center tarda en contestar
        time_contestar = time_conversion_minutos - time_navegacion_minutos
        
        diferencia_horas_list.append(time_contestar)

    suma_time = 0
    for time in diferencia_horas_list:
        suma_time = suma_time + int(time)

    media_time = suma_time / len(call_clean_dataframe)

    print("LA MEDIA DE TIEMPO QUE EL CALL CENTER TARDA EN CONTESTAR ES DE: " + str(media_time/60) + ' HORAS Y ' + str(media_time%60) + ' MINUTOS' )
    
call_clean_dataframe = clean_dataframe.loc[clean_dataframe['lead_type']==' FORM']
get_media_tiempo(call_clean_dataframe)

## 3. Identificar el número de clientes que se pueden localizar por medio de CALL

call_users_dataframe = clean_dataframe.loc[clean_dataframe['lead_type']==' CALL']   
call_users_dataframe

result_list = call_users_dataframe['result'].tolist()
print(result_list[1])

call_users_dataframe = clean_dataframe.loc[clean_dataframe['lead_type']==' CALL'] 
localizable_call_users_dataframe = call_users_dataframe.loc[call_users_dataframe['result']!=' Ilocalizable']
localizable_call_users_dataframe.to_csv('localizable_call_users.csv', index = False)
localizable_call_users_dataframe


## 4. Identificar la lista de sus productos (coches) teniendo en cuenta el ratio de conversión

def get_products(dataframe):
    URLbase_landing = dataframe["URL_base&landing"]
    pages_list = get_pages_list(URLbase_landing)
    products_list = get_products_list(pages_list)
    
    return products_list
products_list = get_products(clean_dataframe)
print(products_list)

def get_pages_list(URLbase_landing):
    pages_list = []
    for element in URLbase_landing:
        if isinstance(element, str):
            splitted_element = element.split("/")
            if splitted_element[len(splitted_element) - 2] == 'es':
                pages_list.append("Home")
            else:
                pages_list.append(splitted_element[len(splitted_element) - 2])
                
    return pages_list

def get_products_list(pages_list):
    pages_set = list(set(pages_list))
    products = []
    for product in pages_set:
        if 'User' not in product and 'Home' not in product and 'home' not in product:
            products.append(product)
            
    return products
