==> UEMC - Master in AI and Big Data/3. Programming techniques/Activity 1/Proyecto/3.Actividad1-Otros_datos_interes.ipynb <==
import pandas as pd
from datetime import datetime
import collections

clean_dataframe = pd.read_csv("df_navegacion_urlseparated_noduplicates_conversiones_concatenated_withtime.csv")

## 1. Número de visitas recibidas al día

Este punto es tan sencillo como recoger la columna 'day' correspondiente al momento de la navegación de la tabla resultante de la limpieza y mirar sus frecuencia

frecuencia_visitas = clean_dataframe['day'].value_counts()
print(frecuencia_visitas)

## 2. Número de visitas que converten y no convierten

total_visitas = len(clean_dataframe)
visitas_convierten = len(clean_dataframe[clean_dataframe['conversiones'] == 1])
visitas_no_convierten = len(clean_dataframe[clean_dataframe['conversiones'] == 0])
porcentaje_convierten = visitas_convierten/total_visitas * 100
porcentaje_no_convierten = visitas_no_convierten/total_visitas * 100
print("TOTAL VISITAS: " + str(total_visitas))
print("CONVIERTEN " + str(visitas_convierten) + ' VISITAS' + '(' +  str(porcentaje_convierten) + '%)')
print("NO CONVIERTEN " + str(visitas_no_convierten) + ' VISITAS'+ '(' +  str(porcentaje_no_convierten) + '%)')

## 3. Número de visitas de tipo de conversión CALL y FORM

form_dataframe = pd.read_csv("form_contact.csv")
call_dataframe = clean_dataframe[clean_dataframe['lead_type'] == ' CALL']
call_dataframe

call_dataframe = clean_dataframe[clean_dataframe['lead_type'] == ' CALL']
form_dataframe = clean_dataframe[clean_dataframe['lead_type'] == ' FORM']

print("NÚMERO DE CONVERSIONES DE TIPO CALL: " + str(len(call_dataframe)))
print("NÚMERO DE CONVERSIONES DE TIPO FORM: " + str(len(form_dataframe)))

## 4. Porcentaje de usuarios recurrentes sobre el total de usuarios

usuarios_recurrentes = clean_dataframe[clean_dataframe['rec'] == 'rec=true']
usuarios_recurrentes = len(usuarios_recurrentes)

total_usuarios = len(clean_dataframe)
porcentaje_usuarios_recurrentes = usuarios_recurrentes/total_usuarios * 100
                                      
print("EL PORCENTAJE DE USUARIOS RECURRENTES SOBRE EL TOTAL DE USUARIOS ES: " + str(porcentaje_usuarios_recurrentes) + '%')


## 5. Coche más visitado

def get_products(dataframe):
    URLbase_landing = dataframe["URL_base&landing"]
    pages_list = get_pages_list(URLbase_landing)
    counter = collections.Counter(pages_list)
    
    return counter
products_list = get_products(clean_dataframe)


def get_pages_list(URLbase_landing):
    pages_list = []
    for element in URLbase_landing:
        if isinstance(element, str):
            splitted_element = element.split("/")
            if splitted_element[len(splitted_element) - 2] == 'es':
                pages_list.append("Home")
            else:
                pages_list.append(splitted_element[len(splitted_element) - 2])
                
    return pages_list

print(products_list)
